# backend/docs/openapi_skills_api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skills API",
    "description": "Recruiter-facing REST endpoints for skill matching, candidate ranking, and skills gap analysis",
    "version": "1.0",
    "contact": {
      "name": "Development Team",
      "email": "support@teamitserve.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "paths": {
    "/api/v1/skills/match": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Match Candidate Skills",
        "description": "Match candidate skills to job requirements.\n\nRequest body:\n{\n    \"candidate_id\": \"uuid\",\n    \"job_id\": \"uuid\"\n}\n\nReturns:\n    Detailed skill match analysis with scoring breakdown",
        "operationId": "match_candidate_skills_api_v1_skills_match_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Match Candidate Skills Api V1 Skills Match Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/skills/rank": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Rank Candidates By Skills",
        "description": "Rank candidates by skill match for a specific job.\n\nRequest body:\n{\n    \"job_id\": \"uuid\",\n    \"limit\": 10,\n    \"min_score\": 0.6\n}\n\nReturns:\n    Ranked list of candidates with skill match scores",
        "operationId": "rank_candidates_by_skills_api_v1_skills_rank_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Rank Candidates By Skills Api V1 Skills Rank Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/skills/gaps": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Analyze Skills Gaps",
        "description": "Analyze skills gaps between candidate and job requirements.\n\nRequest body:\n{\n    \"candidate_id\": \"uuid\",\n    \"job_id\": \"uuid\"\n}\n\nReturns:\n    Detailed gap analysis with prioritized missing skills",
        "operationId": "analyze_skills_gaps_api_v1_skills_gaps_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Analyze Skills Gaps Api V1 Skills Gaps Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_resume_api_v1_candidates_upload_resume_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_api_v1_candidates_upload_resume_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobDescriptionCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "required_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Skills",
            "default": []
          },
          "nice_to_have_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Nice To Have Skills",
            "default": []
          },
          "experience_required": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Required"
          },
          "job_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Type"
          },
          "salary_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Min"
          },
          "salary_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Max"
          }
        },
        "type": "object",
        "required": [
          "title",
          "company",
          "description"
        ],
        "title": "JobDescriptionCreate",
        "description": "Schema for creating job description"
      },
      "JobDescriptionUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "required_skills": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Skills"
          },
          "nice_to_have_skills": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nice To Have Skills"
          },
          "experience_required": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Required"
          },
          "job_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Type"
          },
          "salary_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Min"
          },
          "salary_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Max"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "JobDescriptionUpdate",
        "description": "Schema for updating job description"
      },
      "RefreshTokenRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshTokenRequest",
        "description": "Refresh token request"
      },
      "UpdateUserRoleRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "UpdateUserRoleRequest",
        "description": "Update user role request"
      },
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin",
        "description": "User login schema"
      },
      "UserRegister": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserRegister",
        "description": "User registration schema"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.teamitserve.com",
      "description": "Production server"
    }
  ]
}